// src/app/services/models/api-response-transaction-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { TransactionResponseDto } from '../models/transaction-response-dto';
export interface ApiResponseTransactionResponseDto {
  data?: TransactionResponseDto;
  message?: string;
  success?: boolean;
}